Refactor : 태그 입력 시 마크 찌그러지는 현상 수정
