+++ public/javascript/newsfeed.grouptimeline.js
@@ -25,30 +25,11 @@ function readnewsfeedgrouptimeline(page){
   })
 }
 
+// 무한 스크롤
 async function limitscroll() {
   page++
   readnewsfeedgrouptimeline(page)
   }
-  
-  function debounce(callback, limit = 500) {
-    let timeout;
-    return function (...args) {
-      clearTimeout(timeout);
-      timeout = setTimeout(() => {
-        callback.apply(this, args);
-      }, limit);
-    };
-  }
-  
-  document.addEventListener(
-    'scroll',
-    debounce((e) => {
-      const { clientHeight, scrollTop, scrollHeight } = e.target.scrollingElement;
-      if (clientHeight + scrollTop >= scrollHeight) {
-        limitscroll();
-      }
-    }, 500),
-  );
 
 // 특정 그룹에서 뉴스피드 태그로 검색하기
 function serchtag(tag) {
@@ -175,10 +156,6 @@ async function getTags() {
     }
   });
    let tagHtml = `
-   </div>
-   <svg class="quick-post-footer-action-icon icon-tags">
-   <use xlink:href="#svg-tags"></use>
-    </svg>
    <div class="tag-list">
     ${tags.value.map(tag => `
     <a class="tag-item secondary">${tag}</a>
