+++ views/group-timeline.hbs
@@ -28,7 +28,7 @@ https://cdn.jsdelivr.net/npm/sweetalert2@11.7.3/dist/sweetalert2.min.css
     .newsfeed-image {
   display: flex;
   gap: 1rem;
-  width: 200px;
+  width: 100px;
   height: 80px;
 }
   </style>
@@ -159,7 +159,7 @@ https://cdn.jsdelivr.net/npm/sweetalert2@11.7.3/dist/sweetalert2.min.css
               </div>
               <!-- /QUICK POST FOOTER ACTION -->
               <!-- QUICK POST FOOTER ACTION -->
-              <div class="quick-post-footer-action text-tooltip-tft-medium tag_box" id="tag_box" data-title="태그 추가" onclick="getTags()">
+              <div class="quick-post-footer-action text-tooltip-tft-medium" data-title="태그 추가" onclick="getTags()">
                 <!-- QUICK POST FOOTER ACTION ICON -->
                 <svg class="quick-post-footer-action-icon icon-tags">
                   <use xlink:href="#svg-tags"></use>
@@ -169,6 +169,9 @@ https://cdn.jsdelivr.net/npm/sweetalert2@11.7.3/dist/sweetalert2.min.css
               <!-- /QUICK POST FOOTER ACTION -->
             </div>
             <!-- /QUICK POST FOOTER ACTIONS -->
+            <div class="tag_box" id="tag_box">
+
+            </div>
               <!-- QUICK POST FOOTER ACTIONS -->
               <div class="quick-post-footer-actions">
                 <!-- BUTTON -->
